More work on manuscript figures, including testing different number of varicosities.

+++ bg_net2/bg_net_v2.jl
@@ -12,7 +12,7 @@ mutable struct BgNet
 end
 
 function BgNet(size::Integer, readout_size::Integer, eta_snr::Float64, eta_str::Float64;
-               lambda=0.1, SynapseType::Type=EligabilitySynapse, n_varicosities=10)
+               lambda=0.1, SynapseType::Type=EligabilitySynapse, n_varicosities::Integer=10)
     populations = Dict{Symbol, Population}()
     
     populations[:ctx_exc] = Population(floor(Int, 0.8*size), tau=10.0, noise=0.0)
